Removed TileLayerMetadata and moved it to TileLayerMethods
